+++ src/main/java/Backend/teampple/domain/auth/mapper/RequestOAuthTokenMapper.java
@@ -1,11 +1,11 @@
 package Backend.teampple.domain.auth.mapper;
 
-import Backend.teampple.domain.auth.dto.RequestSignUpDto;
 import Backend.teampple.domain.auth.dto.request.RequestOAuthTokenDto;
+import Backend.teampple.domain.auth.dto.request.RequestSignUpDto;
 import Backend.teampple.global.common.mapper.GenericMapper;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
 
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
-public interface PostOAuthTokenMapper extends GenericMapper<RequestSignUpDto, RequestOAuthTokenDto> {
+public interface RequestOAuthTokenMapper extends GenericMapper<RequestSignUpDto, RequestOAuthTokenDto> {
 }
